streamlit/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
altair==4.2.0
    # via streamlit
appnope==0.1.3
    # via
    #   ipykernel
    #   ipython
asttokens==2.0.8
    # via stack-data
attrs==22.1.0
    # via jsonschema
backcall==0.2.0
    # via ipython
blinker==1.5
    # via streamlit
cachetools==5.2.0
    # via streamlit
certifi==2022.9.14
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via streamlit
commonmark==0.9.1
    # via rich
debugpy==1.6.3
    # via ipykernel
decorator==5.1.1
    # via
    #   ipython
    #   validators
entrypoints==0.4
    # via
    #   altair
    #   jupyter-client
executing==1.0.0
    # via stack-data
gitdb==4.0.9
    # via gitpython
gitpython==3.1.27
    # via streamlit
idna==3.4
    # via requests
importlib-metadata==4.12.0
    # via streamlit
ipykernel==6.15.3
    # via
    #   ipywidgets
    #   pydeck
ipython==8.5.0
    # via
    #   ipykernel
    #   ipywidgets
ipywidgets==8.0.2
    # via pydeck
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
jsonschema==4.16.0
    # via altair
jupyter-client==7.3.5
    # via ipykernel
jupyter-core==4.11.1
    # via jupyter-client
jupyterlab-widgets==3.0.3
    # via ipywidgets
markupsafe==2.1.1
    # via jinja2
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
nest-asyncio==1.5.5
    # via
    #   ipykernel
    #   jupyter-client
numpy==1.23.3
    # via
    #   -r requirements.in
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
packaging==21.3
    # via
    #   ipykernel
    #   streamlit
pandas==1.4.4
    # via
    #   -r requirements.in
    #   altair
    #   streamlit
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.2.0
    # via streamlit
prompt-toolkit==3.0.31
    # via ipython
protobuf==3.20.2
    # via streamlit
psutil==5.9.2
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyarrow==9.0.0
    # via streamlit
pydeck==0.7.1
    # via streamlit
pygments==2.13.0
    # via
    #   ipython
    #   rich
pympler==1.0.1
    # via streamlit
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   pandas
    #   streamlit
pytz==2022.2.1
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyzmq==23.2.1
    # via
    #   ipykernel
    #   jupyter-client
requests==2.28.1
    # via streamlit
rich==12.5.1
    # via streamlit
semver==2.13.0
    # via streamlit
six==1.16.0
    # via python-dateutil
smmap==5.0.0
    # via gitdb
stack-data==0.5.0
    # via ipython
streamlit==1.12.2
    # via -r requirements.in
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
tornado==6.2
    # via
    #   ipykernel
    #   jupyter-client
    #   streamlit
traitlets==5.4.0
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   pydeck
typing-extensions==4.3.0
    # via streamlit
tzdata==2022.2
    # via pytz-deprecation-shim
tzlocal==4.2
    # via streamlit
urllib3==1.26.12
    # via requests
validators==0.20.0
    # via streamlit
wcwidth==0.2.5
    # via prompt-toolkit
widgetsnbextension==4.0.3
    # via ipywidgets
zipp==3.8.1
    # via importlib-metadata
